Extract shared contacts-export core + reusable download button
Factor the participant-contact select, flatten, columns, name-sort pivot
(buildContactsTable), and filename slug helper out of lib/cycle/contacts.ts
into lib/export/contacts.ts so every group export (pod, lab, project, admin
lists) shares one definition. Add ContactsDownloadButton (server component)
and use it on the cycle People tab. Cycle export output is unchanged; its
unit test still passes.
