+++ app/(dashboard)/admin/cycles/[cycle_id]/page.tsx
@@ -5,6 +5,7 @@ import { requireAdmin } from "@/lib/auth/guards";
 import { can, isOwner } from "@/lib/auth/roles";
 import { StatCard, StatusBadge } from "@/app/components/ui";
 import OwnerLifecycle from "@/app/components/owner-lifecycle";
+import { ContactsDownloadButton } from "@/app/components/contacts-download-button";
 import CycleStatusForm from "./cycle-status-form";
 import {
   CycleScheduleForm,
@@ -487,13 +488,9 @@ export default async function AdminCycleDetailPage({
             {isOrg ? "Core contributors" : "Participants"} (
             {participants.length})
           </h2>
-          <a
+          <ContactsDownloadButton
             href={`/api/cycles/${cycleId}/contacts/export`}
-            download
-            className="inline-flex items-center gap-1.5 rounded-card bg-teal-deep px-4 py-2 text-sm font-semibold tracking-tight text-white transition-colors duration-150 hover:bg-teal focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-teal focus-visible:ring-offset-2"
-          >
-            Download contacts CSV
-          </a>
+          />
         </div>
         <ParticipantsTable
           participants={participants}
